AddedAdmin

+++ app/api/adminDashBoard/books/route.ts
@@ -3,7 +3,8 @@
 import { prisma } from "@/lib/prisma";
 import { NextRequest, NextResponse } from "next/server";
 import { string } from "zod";
-
+import fs from "fs";
+import path from "path";
 export async function POST(req: Request) {
   const { title, author, pages, price, image } = await req.json();
   if (!title || !author || !pages || !price) {
@@ -38,8 +39,27 @@ export async function DELETE(req: NextRequest) {
   if (!isExist) {
     return NextResponse.json({ error: "Book Doesn't Exist!" });
   }
-  await prisma.books.delete({ where: { id } });
-  return NextResponse.json({
-    message: "Book Deleted successfully!",
-  });
+  const filePath = path.join(
+    process.cwd(),
+    "public",
+    "uploads",
+    isExist.image as string
+  );
+  try {
+    if (fs.existsSync(filePath)) {
+      fs.unlinkSync(filePath); // Delete the file
+      await prisma.books.delete({ where: { id } });
+      return NextResponse.json({
+        message: "Book Deleted successfully!",
+      });
+    } else {
+      return NextResponse.json({ error: "File not found" }, { status: 404 });
+    }
+  } catch (error) {
+    console.error("Error deleting file:", error);
+    return NextResponse.json(
+      { error: "Failed to delete file" },
+      { status: 500 }
+    );
+  }
 }
